Enhance exercise 2.29(b).
The original way of solving the problem is long-winded, with unnecessary
repeated statements. The problem is that it didn't follow the principle
"hierarchicy". Binary mobile consists of branches, while branches could
include a mobile. There is no specification that there should be an
empty value in a mobile or branch. So it needn't be considered at the
very beginning. Thus we can program in terms of the definition directly
rather than cover exceptional cases.